refactor(app): use common model for edition links
