+++ src/app/core/navbar/navbar.data.ts
@@ -1,35 +1,14 @@
 import { faEnvelope, faFileAlt, faHome, faNetworkWired } from '@fortawesome/free-solid-svg-icons';
 
-import { EDITION_ROUTE_CONSTANTS } from '@awg-views/edition-view/edition-route-constants';
+import { LabeledRoute } from '@awg-shared/models/labeled-route.model';
+import { EDITION_ROUTE_CONSTANTS } from '@awg-views/edition-view/edition-routes.constants';
 
-import { NavbarDropdownLink, NavbarItems } from './navbar.model';
-
-/**
- * Object constant with a set of navigation items for the navbar.
- *
- * It keeps the navigation items for the navbar.
- */
-export const NAVBAR_DROPDOWN_EDITION_GENERAL_LINKS: NavbarDropdownLink[] = [
-    {
-        label: EDITION_ROUTE_CONSTANTS.SERIES.full,
-        route: [EDITION_ROUTE_CONSTANTS.EDITION.route, EDITION_ROUTE_CONSTANTS.SERIES.route],
-    },
-    {
-        label: EDITION_ROUTE_CONSTANTS.ROWTABLES.full,
-        route: [EDITION_ROUTE_CONSTANTS.EDITION.route, EDITION_ROUTE_CONSTANTS.ROWTABLES.route],
-    },
-    {
-        label: EDITION_ROUTE_CONSTANTS.PREFACE.full,
-        route: [EDITION_ROUTE_CONSTANTS.EDITION.route, EDITION_ROUTE_CONSTANTS.PREFACE.route],
-    },
-];
+import { NavbarItems } from './navbar.model';
 
 /**
  * Object constant with a set of section edition links for the navbar.
- *
- * It keeps the section edition links for the navbar.
  */
-export const NAVBAR_DROPDOWN_EDITION_SECTION_LINKS: NavbarDropdownLink[] = [
+export const NAVBAR_DROPDOWN_EDITION_SECTION_LINKS: LabeledRoute[] = [
     {
         label: 'Einleitung / Intro',
         route: [EDITION_ROUTE_CONSTANTS.EDITION_INTRO.route],
@@ -42,8 +21,6 @@ export const NAVBAR_DROPDOWN_EDITION_SECTION_LINKS: NavbarDropdownLink[] = [
 
 /**
  * Object constant with a set of navigation items for the navbar.
- *
- * It keeps the navigation items for the navbar.
  */
 export const NAVBAR_ITEMS: NavbarItems = {
     home: {
