+++ src/app/views/edition-view/edition-outlets/edition-complex/edition-detail/edition-intro/edition-intro-partial-disclaimer/edition-intro-partial-disclaimer.component.spec.ts
@@ -13,7 +13,7 @@ import {
 } from '@testing/expect-helper';
 import { RouterLinkStubDirective } from '@testing/router-stubs';
 
-import { EDITION_ROUTE_CONSTANTS } from '@awg-views/edition-view/edition-route-constants';
+import { EDITION_ROUTE_CONSTANTS } from '@awg-views/edition-view/edition-routes.constants';
 import { EditionComplex } from '@awg-views/edition-view/models';
 import { EditionComplexesService } from '@awg-views/edition-view/services';
