+++ src/app/views/edition-view/edition-outlets/edition-complex/edition-detail/edition-report/edition-report.component.ts
@@ -6,7 +6,7 @@ import { EMPTY } from 'rxjs';
 import { catchError, switchMap } from 'rxjs/operators';
 
 import { ModalComponent } from '@awg-shared/modal/modal.component';
-import { EDITION_ROUTE_CONSTANTS } from '@awg-views/edition-view/edition-route-constants';
+import { EDITION_ROUTE_CONSTANTS } from '@awg-views/edition-view/edition-routes.constants';
 import {
     EditionComplex,
     SourceDescriptionList,
