+++ src/app/views/edition-view/models/edition-complex.model.ts
@@ -1,7 +1,7 @@
 import { MetaPerson } from '@awg-shared/meta/meta.model';
 import { PERSONS_DATA } from '@awg-shared/meta/persons.data';
 
-import { EDITION_CATALOGUE_TYPE_CONSTANTS, EDITION_ROUTE_CONSTANTS } from '../edition-route-constants';
+import { EDITION_CATALOGUE_TYPE_CONSTANTS, EDITION_ROUTE_CONSTANTS } from '../edition-routes.constants';
 import { EditionRouteConstant } from './edition-route-constant.model';
 
 /**
