+++ src/app/views/edition-view/services/edition-data.service.ts
@@ -5,7 +5,7 @@ import { Observable, forkJoin as observableForkJoin, of as observableOf } from '
 import { catchError, defaultIfEmpty, take } from 'rxjs/operators';
 
 import { EDITION_ASSETS_DATA } from '@awg-views/edition-view/data';
-import { EDITION_ROUTE_CONSTANTS } from '@awg-views/edition-view/edition-route-constants';
+import { EDITION_ROUTE_CONSTANTS } from '@awg-views/edition-view/edition-routes.constants';
 import {
     EditionComplex,
     EditionRowTablesList,
