+++ src/app/views/home-view/home-view.component.spec.ts
@@ -17,7 +17,7 @@ import { HeadingComponent } from '@awg-shared/heading/heading.component';
 import { META_DATA } from '@awg-shared/meta/meta.data';
 import { MetaPage, MetaSectionTypes } from '@awg-shared/meta/meta.model';
 import { ScrollToTopButtonComponent } from '@awg-shared/scroll-to-top-button/scroll-to-top-button.component';
-import { EDITION_ROUTE_CONSTANTS } from '@awg-views/edition-view/edition-route-constants';
+import { EDITION_ROUTE_CONSTANTS } from '@awg-views/edition-view/edition-routes.constants';
 import { EditionOutlineSection, EditionSectionLink } from '@awg-views/edition-view/models';
 import { EditionComplexesService, EditionOutlineService } from '@awg-views/edition-view/services';
 
@@ -91,7 +91,7 @@ describe('HomeViewComponent (DONE)', () => {
     let expectedPageMetaData: MetaPage;
     let expectedSectionLinksData: EditionSectionLink[];
 
-    let expectedRouterlinks: string[][];
+    let expectedRouterLinks: string[][];
 
     beforeAll(() => {
         EditionComplexesService.initializeEditionComplexesList();
@@ -151,7 +151,7 @@ describe('HomeViewComponent (DONE)', () => {
                 separator: '',
             },
         ];
-        expectedRouterlinks = getRouterlinks(expectedSections);
+        expectedRouterLinks = getRouterlinks(expectedSections);
 
         // Create component fixture
         fixture = TestBed.createComponent(HomeViewComponent);
@@ -190,7 +190,7 @@ describe('HomeViewComponent (DONE)', () => {
         it('... should have signal `rowtablesRoute` to hold the provided route', () => {
             expectToBe(isSignal(component.rowtablesRoute), true);
 
-            expectToEqual(component.rowtablesRoute(), expectedRouterlinks.at(-2));
+            expectToEqual(component.rowtablesRoute(), expectedRouterLinks.at(-2));
         });
 
         describe('VIEW', () => {
@@ -413,22 +413,22 @@ describe('HomeViewComponent (DONE)', () => {
                 linkDes = getAndExpectDebugElementByDirective(
                     compDe,
                     RouterLink,
-                    expectedRouterlinks.length,
-                    expectedRouterlinks.length
+                    expectedRouterLinks.length,
+                    expectedRouterLinks.length
                 );
 
                 routerLinks = linkDes.map(de => de.injector.get(RouterLink) as RouterLink);
             });
 
             it('... can get correct number of routerLinks from template', () => {
-                expectToBe(routerLinks.length, expectedRouterlinks.length);
+                expectToBe(routerLinks.length, expectedRouterLinks.length);
             });
 
             it('... can get correct linkParams from template', () => {
                 for (const [index, routerLink] of routerLinks.entries()) {
                     const urlTree = routerLink.urlTree;
 
-                    expectToBe(urlTree.toString(), expectedRouterlinks[index].join('/'));
+                    expectToBe(urlTree.toString(), expectedRouterLinks[index].join('/'));
                 }
             });
 
@@ -439,7 +439,7 @@ describe('HomeViewComponent (DONE)', () => {
                     navigateSpy.mockClear();
 
                     const linkDe = linkDes[index];
-                    const expectedRouterLink = expectedRouterlinks[index];
+                    const expectedRouterLink = expectedRouterLinks[index];
 
                     await clickAndAwaitChanges(linkDe, fixture);
